+++ darwin/onnxruntime/Sources/onnxruntime_ffi/include/onnxruntime_ffi.h
@@ -1,0 +1,8 @@
+#ifndef ONNXRUNTIME_FLUTTER_FFI_H
+#define ONNXRUNTIME_FLUTTER_FFI_H
+
+#include <onnxruntime/onnxruntime_c_api.h>
+#include <onnxruntime/cpu_provider_factory.h>
+#include <onnxruntime/coreml_provider_factory.h>
+
+#endif